Add export functionality, cleanup
