+++ package.json
@@ -29,6 +29,7 @@
         "@mantine/hooks": "^4.0.7",
         "@mantine/next": "^4.0.7",
         "axios": "^0.26.1",
+        "file-saver": "^2.0.5",
         "html-to-image": "^1.9.0",
         "next": "^12.1.0",
         "react": "^17.0.2",
